+++ CHANGELOG.md
@@ -11,6 +11,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add missing string for "full sponsored" support grant type in the Conversions
   CSV export
 
+### Added
+
+- Add more columns to the Academies due to transfer export: School phase, school
+  age range, main contact details, school type
+
 ## [Release-52][release-52]
 
 ### Changed